Extend MPI support to Cray compilers
This commit changes the MPI search function to support simple compiler
and test detection.

The advantages of this approach is that one can select the list of
compilers to support in a single place, instead of multiple if else
combinations. Furthermore it also tests whether the actual compiler
works by testing it against a simple mpi example program.