Add the @ workdir syntax
This introduces the `[@ workdir @]` syntax. In this syntax the users can
change dir just before the execution of a given task.
The feature may seem strange at first because users can easily add tasks
that have `cd` commands before the executable call. However, the `cd`
command does not work for basic engine combined with the srun method.
Therefore it is necessary to introduce such feature.

All engines and run methods implement this feature, therefore it is easy
for the end user to migrate task files between greasy installation that
use different engines.

Additionally, the logic of emitting `srun` commands for the basic engine
is slightly different from the original code. In this implementation one
always emit the srun command even when executing tasks in the master
node. This is due to the fact that SLURM allows for SPANK plugins to be
defined for each job step (`srun` call). In order to avoid that some
tasks are executed as job steps and others not, then one needs to always
invoke thes tasks using the `srun` command.